Fix windows warning by initializing to double

+++ rosbag2_transport/src/rosbag2_transport/player.cpp
@@ -216,7 +216,7 @@ void Player::prepare_publishers(const PlayOptions & options)
 
 void Player::prepare_clock(const PlayOptions & options, rcutils_time_point_value_t starting_time)
 {
-  float rate = options.rate > 0.0 ? options.rate : 1.0;
+  double rate = options.rate > 0.0 ? options.rate : 1.0;
   clock_ = std::make_unique<rosbag2_cpp::TimeControllerClock>(starting_time, rate);
 }
